fix(compute-env): state upgrader for azure_batch delete_jobs_on_completion
Reported in v0.40.0-rc1 internal test against tf-seqera-showcase: users
upgrading from v0.30.x with `azure_batch.delete_jobs_on_completion =
"on_success"` in state hit a spurious `null -> true` diff on the new
`delete_jobs_on_completion_enabled` field after they update their config,
forcing a resource replacement.

Fix:

  - Bump `seqera_compute_env` schema entity version 1 -> 2 (overlays/compute-env.yaml).
  - Add `ComputeenvStateUpgraderV1` (v1 -> v2 upgrader) in
    internal/stateupgraders/computeenv_v1.go that walks `config.azure_batch`
    and sets `delete_jobs_on_completion_enabled = true` whenever the old
    string had a non-empty value. Doesn't overwrite an explicit user-set
    value. Other platforms are untouched.
  - Drop the dead `compute_env_id -> id` rename function that previously
    sat unused in computeenv_v0.go.

CHANGELOG entry under v0.40.0 BREAKING CHANGES with the diff/migration
guidance the user requested.

Verified: `go build ./...`, `go vet ./...`, `golangci-lint run ./...` all
clean. Schema version in `internal/provider/computeenv_resource.go` is
now 2 with both V0 and V1 upgraders wired.
