+++ .speakeasy/gen.lock
@@ -1,7 +1,7 @@
 lockVersion: 2.0.0
 id: 9eba9f63-fbb7-47c2-b957-92fb58cab346
 management:
-  docChecksum: 3624fa8fdd73e50a486be4704b356329
+  docChecksum: 56047aa0e15176568d7f007127162de0
   docVersion: 1.147.0
   speakeasyVersion: 1.761.9
   generationVersion: 2.881.4
@@ -489,7 +489,7 @@ trackedFiles:
     pristine_git_object: 0c1b43d5db49b023cf3c7408c198a4d66c795d8a
   internal/provider/computeenv_resource.go:
     id: f4162f219efd
-    last_write_checksum: sha1:1b1f21fa91f6781ba0d1b049b4baadf5b177fd8c
+    last_write_checksum: sha1:d1ab9a793c325aef7f709978c9983d258da133bb
     pristine_git_object: d5550a0a3bc980d281c55531d71458e7e4968b48
   internal/provider/computeenv_resource_sdk.go:
     id: 1e29eea44332
