+++ .speakeasy/out.openapi.yaml
@@ -14577,7 +14577,7 @@ components:
           type: integer
           format: int64
       x-speakeasy-entity: ComputeEnv
-      x-speakeasy-entity-version: 1
+      x-speakeasy-entity-version: 2
       x-speakeasy-entity-description: |
         Seqera Platform compute environments define the execution platform where a pipeline will run.
         Compute environments enable users to launch pipelines on a growing number of cloud and
