+++ .speakeasy/workflow.lock
@@ -2,8 +2,8 @@ speakeasyVersion: 1.761.9
 sources:
     Seqera API:
         sourceNamespace: seqera-api
-        sourceRevisionDigest: sha256:2e8a0762e3e04f5cd154d25613eee21ad4cfbc24b1b1889d04d8c42e770fdb37
-        sourceBlobDigest: sha256:79b97356ed6c125bb1a834e90510ed51077fe079c9e4393f1f8c239c6c3c7146
+        sourceRevisionDigest: sha256:577cbe032ffdb43850ee3f456a30f7be4fba62bbeb7be5aeb01a299c999bc738
+        sourceBlobDigest: sha256:c1a2f008231b60082d0d045058329a029a1991133e0a3af09ab922d421ce49a3
         tags:
             - latest
             - 1.147.0
@@ -11,8 +11,8 @@ targets:
     seqera:
         source: Seqera API
         sourceNamespace: seqera-api
-        sourceRevisionDigest: sha256:2e8a0762e3e04f5cd154d25613eee21ad4cfbc24b1b1889d04d8c42e770fdb37
-        sourceBlobDigest: sha256:79b97356ed6c125bb1a834e90510ed51077fe079c9e4393f1f8c239c6c3c7146
+        sourceRevisionDigest: sha256:577cbe032ffdb43850ee3f456a30f7be4fba62bbeb7be5aeb01a299c999bc738
+        sourceBlobDigest: sha256:c1a2f008231b60082d0d045058329a029a1991133e0a3af09ab922d421ce49a3
 workflow:
     workflowVersion: 1.0.0
     speakeasyVersion: latest
